requirements.txt
# Telegram Bot per Affiliate Marketing
# Python 3.8+ richiesto

python-telegram-bot==20.7
selenium==4.16.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
lxml==5.1.0
python-dotenv==1.0.0
